Made all cycle counters int64_t.
Also reduced the status of most functions in CommandAnalysis.cc from public to private, and made 3 variables local to a function instead of them being a member variable.

+++ src/MemoryPowerModel.h
@@ -134,7 +134,7 @@ class MemoryPowerModel {
                           double spup_ref_pre_cycles,
                           double clk);
 
-  double total_cycles;
+  int64_t total_cycles;
 
   struct Energy {
     // Total energy of all activates
